requirements.txt
# Core dependencies
requests>=2.28.0

# LLM providers (optional - install as needed)
openai>=1.0.0  # For OpenAI GPT models and Azure OpenAI
anthropic>=0.8.0  # For Anthropic Claude models
google-generativeai>=0.3.0  # For Google Gemini models
cohere>=4.21.0  # For Cohere models
boto3>=1.26.0  # For AWS Bedrock models
together>=0.2.6  # For Together AI models
replicate>=0.15.0  # For Replicate models

# Local LLM support (optional)
transformers>=4.21.0  # For Hugging Face models
torch>=1.12.0  # For Hugging Face models
# Note: Ollama support uses HTTP requests, no additional packages needed

# For examples and development
python-dotenv>=0.19.0  # For environment variable management 